🐛 fix: category 모델 하나의 파일로 병합 및 related_name 추가, 순환참조 해결을 위한 참조 문자열로 변환
