--- apps/lectures/models/lecture_category.py
@@ -1,18 +1,0 @@
-from django.db import models
-
-from apps.core.models import TimeStampedModel
-from apps.lectures.models import CrawledLecture
-from apps.lectures.models.category import Category
-
-
-class LectureCategory(TimeStampedModel):
-    pk = models.CompositePrimaryKey("lecture_id", "category_id", help_text="복합키")
-    lecture = models.ForeignKey(CrawledLecture, on_delete=models.CASCADE, help_text="강의 id")
-    category = models.ForeignKey(Category, on_delete=models.CASCADE, help_text="카테고리 id")
-
-    class Meta:
-        ordering = ["pk"]
-        db_table = "lecture_categories"
-
-    def __str__(self) -> str:
-        return f"[{self.lecture}] {self.category}"
